requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.24.3
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.29.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.4
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.5.0
